+++ src/js/api/streams.js
@@ -6,7 +6,6 @@
 
 import browser from 'webextension-polyfill'
 import { fromPromise, merge } from 'kefir'
-import { path, prop } from 'rambda'
 
 import { browserEvent$ } from './helpers'
 
@@ -15,20 +14,12 @@ export const getTree$ = () => {
 }
 
 export const bookmarkCreated$ = browserEvent$(browser.bookmarks.onCreated)
-    .spy('[api/streams] Bookmark created:')
-
 export const bookmarkRemoved$ = browserEvent$(browser.bookmarks.onRemoved)
-    .spy('[api/streams] Bookmark removed:')
+// export const bookmarkChanged$ = browserEvent$(browser.bookmarks.onChanged)
+// export const bookmarkMoved$ = browserEvent$(browser.bookmarks.onMoved)
 
-export const bookmarkChanged$ = browserEvent$(browser.bookmarks.onChanged)
-    .spy('[api/streams] Bookmark changed:')
-
-export const bookmarkMoved$ = browserEvent$(browser.bookmarks.onMoved)
-    .spy('[api/streams] Bookmark moved:')
-
-export const bookmarksModified$ = merge([
-    bookmarkChanged$.map(prop(1)),
-    bookmarkCreated$.map(prop(1)),
-    bookmarkMoved$.map(prop(1)),
-    bookmarkRemoved$.map(path([1, 'node'])),
+export const bookmarkCountChanged$ = merge([
+    bookmarkCreated$,
+    bookmarkRemoved$,
 ])
+    .debounce(250, { immediate: false })