+++ src/js/components/Popup.svelte
@@ -3,8 +3,9 @@
     import { indexBy, prop, sortBy, toPairs } from 'rambda'
     import { onDestroy, onMount } from 'svelte'
 
-    import { getCurrentTab } from '../api/tabs'
+    import { activeTabQuery } from '../api/tabs'
     import { messages } from '../lib/messaging'
+    import { bookmarkCountChanged$ } from '../api/streams'
     import { t } from '../lib/translate'
 
     import { currentTab } from '../stores/currentTab'
@@ -21,14 +22,20 @@
             : t('add_bookmark')
     )
 
+    async function currentTabStatus () {
+        const tabs = await browser.tabs.query(activeTabQuery)
+        const tab = tabs[0]
+
+        // console.debug('[Popup] currentTabStatus:', tab)
+        if (tab) {
+            $currentTab = { ...tab }
+            messages.emit('api', { action: 'savedBookmarks', tab })
+        }
+    }
+
     async function deleteBookmark (bookmark) {
         // console.debug('Deleting bookmark:', bookmark)
         browser.bookmarks.remove(bookmark.id)
-            .then(() => {
-                // console.info('Bookmark deleted:', bookmark)
-                location.reload()
-            })
-            .then(updateStatus)
             .catch(console.error)
     }
 
@@ -37,35 +44,29 @@
         return false
     }
 
-    function updateStatus () {
-        getCurrentTab().then(tab => {
-            // console.debug('[Popup] updating bookmark status:', { tab })
-            $currentTab = { ...tab }
-            messages.emit('api', { action: 'bookmarkStatus', tab })
-        })
-    }
-
-    function updateBookmarks (bookmarks) {
+    function updateSavedBookmarks (bookmarks) {
         const sorted = sortBy(prop('dateAdded'), bookmarks || [])
         const saved = new Map(toPairs(indexBy(prop('id'), sorted)))
-        // console.debug('[Popup] updateBookmarks sorted:', { sorted, saved })
+        // console.debug('[Popup] updateBookmarks sorted:', { bookmarks, sorted, saved })
 
-        // TODO Use separate store?
         $savedBookmarks = saved
     }
 
     onMount(() => {
-        messages.on('bookmarkStatus', updateBookmarks)
+        messages.on('savedBookmarks', updateSavedBookmarks)
         messages.on('button:close', onClose)
         messages.on('deleteBookmark', deleteBookmark)
 
+        currentTabStatus()
+
         messages.emit('api', { action: 'recentCategories' })
 
-        updateStatus()
+        // Refresh contents when bookmarks change
+        bookmarkCountChanged$.observe(currentTabStatus)
     })
 
     onDestroy(() => {
-        messages.off('bookmarkStatus', updateBookmarks)
+        messages.off('savedBookmarks', updateSavedBookmarks)
         messages.off('button:close', onClose)
         messages.off('deleteBookmark', deleteBookmark)
     })